+++ src/main/java/com/springboot/monew/common/entity/BaseUpdatableEntity.java
@@ -1,10 +1,8 @@
-package com.springboot.monew.base;
+package com.springboot.monew.common.entity;
 
-import jakarta.persistence.EntityListeners;
 import jakarta.persistence.MappedSuperclass;
 import lombok.Getter;
 import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import java.time.Instant;
